Move path utils to stepper
