+++ client/landing/stepper/utils/path.ts
